# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "App",
        "version": "0.1"
    },
    "paths": {
...
        "/api/cms/products": {
            "get": {
                "tags": [
                    "CMS Products"
                ],
                "summary": "Lấy ra danh sách sản phẩm",
                "operationId": "dbc577ad22a0d09f0afb023c7763e302",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Trang",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "perpage",
                        "in": "query",
                        "description": "Per Page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "searchFields[]",
                        "in": "query",
                        "description": "List of fields to search. Example: ['name_booking']",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "example": "name_booking"
                            }
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Content search. Example: 'Thiên'",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Thiên"
                        }
                    },
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "Filter criteria in JSON format. Example: {\"created_at_RANGE\": [\"2024-01-20\", \"2024-01-28\"]}",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "{\"created_at_RANGE\": [\"2024-01-20\", \"2024-01-28\"]}"
                        }
                    },
                    {
                        "name": "sort[]",
                        "in": "query",
                        "description": "Sort criteria in array format. Use '-' for descending order and '+' for ascending order. Example: ['-created_at']",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "example": "-created_at"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Success."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cms/products/{id}": {
            "get": {
                "tags": [
                    "CMS Products"
                ],
                "summary": "Lấy thông tin chi tiết sản phẩm",
                "operationId": "75cd0a156b2ad748b3fa1aaaace85888",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của sản phẩm",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A product detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cms/products/create": {
            "post": {
                "tags": [
                    "CMS Products"
                ],
                "summary": "Tạo sản phẩm mới",
                "operationId": "6a4881953f7193dffa4d7b9aa3ba8934",
                "parameters": [
                    {
                        "name": "X-Localization",
                        "in": "header",
                        "description": "Ngôn ngữ",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "vi"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cms/products/update/{id}": {
            "post": {
                "tags": [
                    "CMS Products"
                ],
                "summary": "Cập nhật sản phẩm",
                "operationId": "6ff67df17a82dca828e3469a77b6d594",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "_method",
                                    "name",
                                    "price",
                                    "category_id"
                                ],
                                "properties": {
                                    "_method": {
                                        "type": "string",
                                        "example": "PUT"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Một vài thông tin về "
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "..."
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 40000
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật thành công"
                    },
                    "422": {
                        "description": "Lỗi validate"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cms/products/delete/{id}": {
            "delete": {
                "tags": [
                    "CMS Products"
                ],
                "summary": "Xóa sản phẩm",
                "description": "Xóa sản phẩm khỏi hệ thống",
                "operationId": "b957b4f9c35b5ea0e59ef5a4731f9d88",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của sản phẩm",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sản phẩm đã được xoá",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sản phẩm đã được xoá thành công"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy sản phẩm"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cms/products/import-excel": {
            "post": {
                "tags": [
                    "CMS Products"
                ],
                "summary": "Import danh sách sản phẩm từ file CSV",
                "operationId": "3673d1d0a6d3d8cc3ff9ae60cc116ef3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "file": {
                                        "description": "File CSV chứa danh sách sản phẩm",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Import thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Import thành công!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Lỗi validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "File không hợp lệ"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Product": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number",
                        "format": "float"
                    },
                    "category_id": {
                        "type": "integer"
                    },
                    "image": {
                        "type": "string",
                        "format": "binary"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}